feat(rename): rename package to aviod duplication in pypi
